[REFACTOR] Reorganize style utils code
Switch `MarkerIdentifier` from enum to class: there is no need for an enum to
store constants that don't need to be enumerated.
Move identifiers and utils to dedicated file in the new 'style' subdirectory.
Simplify imports and exports related to style utils.
