+++ src/component/mxgraph/config/MarkerConfigurator.ts
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-import { MarkerIdentifier } from '../StyleUtils';
+import { MarkerIdentifier } from '../style';
 import { mxgraph } from '../initializer';
 import { mxAbstractCanvas2D, mxCell, mxPoint, mxShape } from 'mxgraph'; // for types
