+++ src/component/mxgraph/renderer/StyleComputer.ts
@@ -27,7 +27,7 @@ import {
   ShapeBpmnStartEvent,
   ShapeBpmnSubProcess,
 } from '../../../model/bpmn/internal/shape/ShapeBpmnElement';
-import { BpmnStyleIdentifier } from '../StyleUtils';
+import { BpmnStyleIdentifier } from '../style';
 import { ShapeBpmnCallActivityKind, ShapeBpmnElementKind, ShapeBpmnMarkerKind, ShapeUtil } from '../../../model/bpmn/internal';
 import { AssociationFlow, SequenceFlow } from '../../../model/bpmn/internal/edge/flows';
 import { Font } from '../../../model/bpmn/internal/Label';
