+++ src/component/mxgraph/renderer/style-utils.ts
