+++ src/component/mxgraph/shape/event-shapes.ts
@@ -17,7 +17,7 @@
 import { ShapeBpmnEventDefinitionKind } from '../../../model/bpmn/internal';
 import { BpmnCanvas, PaintParameter, IconPainterProvider } from './render';
 import { buildPaintParameter } from './render/icon-painter';
-import StyleUtils, { StyleDefault } from '../StyleUtils';
+import { StyleDefault, StyleUtils } from '../style';
 import { mxAbstractCanvas2D } from 'mxgraph';
 import { mxgraph } from '../initializer'; // for types
