+++ src/component/mxgraph/shape/flow-shapes.ts
@@ -16,7 +16,7 @@
 
 import { IconPainterProvider } from './render';
 import { buildPaintParameter } from './render/icon-painter';
-import StyleUtils from '../StyleUtils';
+import { StyleUtils } from '../style';
 import { MessageVisibleKind } from '../../../model/bpmn/internal/edge/kinds';
 import { mxgraph } from '../initializer';
 import { mxAbstractCanvas2D, mxRectangle } from 'mxgraph'; // for types
