+++ src/component/mxgraph/shape/render/icon-painter.ts
@@ -15,7 +15,7 @@
  */
 
 import { BpmnCanvas } from './BpmnCanvas';
-import StyleUtils from '../../StyleUtils';
+import { StyleUtils } from '../../style';
 import { IconStyleConfiguration, ShapeConfiguration, Size } from './render-types';
 import { mxAbstractCanvas2D, mxShape } from 'mxgraph'; // for types
